% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDMA_0.1_functions_analyses_2025-11-06.r
\name{calcFM}
\alias{calcFM}
\title{Calculate Form Matrix (FM)}
\usage{
calcFM(x)
}
\arguments{
\item{x}{A matrix with landmarks as rows and coordinates (X, Y, Y) as columns.}
}
\value{
A pairwise distance matrix (of class \code{dist} and \code{FM}) for Euclidean distance 
  between all pairs of landmarks in the data set.  Overall object size is calculated as the 
  geometric mean of all distances and is returned as the attribute \code{size} in the \code{FM} 
  object.
}
\description{
Generate the Euclidean distance form matrix (FM) for a single object.
}
\examples{
FM1 <- calcFM(guenons$rawcoords[,,1])
str(FM1)
plot(FM1)
}
