% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDMA_0.1_functions_analyses_2025-11-11.r
\name{calcSM}
\alias{calcSM}
\title{Calculate Shape Matrix (SM)}
\usage{
calcSM(x)
}
\arguments{
\item{x}{A matrix with landmarks as rows and coordinates (X, Y, Y) as columns.}
}
\value{
A pairwise distance matrix (of class \code{dist} and \code{SM}) for Euclidean distance 
  between all pairs of landmarks in the data set, scaled by size.  Overall object size is 
  calculated as the geometric mean of all distances and is returned as the attribute \code{size} 
  in the \code{SM} object.
}
\description{
Generate the Euclidean distance shape matrix (SM) for a single object.
}
\examples{
SM1 <- calcSM(guenons$rawcoords[,,1])
str(SM1)
plot(SM1)
}
