% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDMA_0.1_functions_analyses_2025-11-08.r
\name{EDMApcoa}
\alias{EDMApcoa}
\title{Principal Coordinates Analysis Ordination for Multiple Objects}
\usage{
EDMApcoa(x)
}
\arguments{
\item{x}{An array of landmarks for multiple specimens, with the first dimension corresponding to 
landmarks, the second corresponding to X, Y, and Z, and the third corresponding to specimens.}
}
\value{
A list return by \code{cmdscale} containing the principal coordinates and associated 
  eigenvalues.
}
\description{
Perform a Principal Coordinates Analysis (PCoA) quantifying shape variation for a set of objects 
  with a common set of landmarks.  Performs \code{cmdscale()} on a pairwise distance matrix 
  for all objects (calculated using \code{shapediff}).
}
\examples{
PCoA <- EDMApcoa(guenons$rawcoords[,,guenons$genus \%in\% c("Erythrocebus", "Miopithecus")])
plot(PCoA$points, pch=21, bg=guenons$genus[guenons$genus \%in\% c("Erythrocebus", "Miopithecus")])
}
