% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_0.1_2025-11-08.r
\docType{data}
\name{guenons}
\alias{guenons}
\title{GueSDat, the free Guenon Skull Database (Cardini & Elton, 2017)}
\format{
...
}
\usage{
guenons
}
\description{
Cranial landmarks for 1,315 cercopithecine crania.
}
\examples{
plot_lmks(guenons$rawcoords[,,1],
     segs=guenons$wireframe_segments,
		tris=guenons$triangles,
		label=TRUE)
}
\references{
Cardini A, Elton S. (2017) Is there a "Wainer's rule"? Testing which sex varies most 
  as an example analysis using GueSDat, the free Guenon Skull Database. 
  \emph{Hystrix, the Italian Journal of Mammalogy}. 28:147-156. 
  (\href{https://doi.org/10.4404/hystrix-28.2-12139}{https://doi.org/10.4404/hystrix-28.2-12139})
}
\keyword{datasets}
