% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDMA_0.1_functions_analyses_2025-11-08.r
\name{influentiallmk}
\alias{influentiallmk}
\title{Visualize Influential Landmarks}
\usage{
influentiallmk(
  A,
  B,
  highlight = NULL,
  pt.bg = "#00000020",
  hl.pch = 4,
  hl.col = "#FF0000FF"
)
}
\arguments{
\item{highlight}{An integer value specifying the number of a landmark to highlight in plots. If 
provided, all logged ratios for interlandmark distances involving that landmark will be highlighted 
in the plot.  Defaults to \code{NULL}.}

\item{pt.bg}{A color value to use in plotting logged ratios.}

\item{hl.pch}{An integer specifying the plotting character to use for highlighted points when a 
value is supplied to \code{highlight}.}

\item{hl.col}{A color value to use for highlighted points when a value is supplied to \code{highlight}.}

\item{x}{An array of landmarks for multiple specimens, with the first dimension corresponding to 
landmarks, the second corresponding to X, Y, and Z, and the third corresponding to specimens.}
}
\description{
Calculates a measure of spatial variability for each landmark in relation to all other landmarks 
  for a pair of objects.  This is (mostly) the visualization method of Cole and Richtsmeier (1998), but 
  it presents the log of ratios as opposed to raw ratios, and shows the ratio from shape matrices as 
  well as form matrices.
}
\examples{
influentiallmk(A=guenons$rawcoords[,,guenons$genus=="Erythrocebus"][,,1],
               B=guenons$rawcoords[,,guenons$genus=="Miopithecus"][,,1],
               highlight=57)
influentiallmk(A=guenons$rawcoords[,,guenons$genus=="Erythrocebus"][,,1],
               B=guenons$rawcoords[,,guenons$genus=="Miopithecus"][,,1],
               highlight=c(57,131), hl.pch=5, hl.col="blue")
}
