% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDMA_0.1_functions_analyses_2025-11-06.r
\name{lmkdisper}
\alias{lmkdisper}
\title{Quantifying Landmark Dispersion}
\usage{
lmkdisper(
  x,
  plt = FALSE,
  highlight = NULL,
  pt.bg = "#00000020",
  hl.pch = 4,
  hl.col = "#FF000040"
)
}
\arguments{
\item{x}{An array of landmarks for multiple specimens, with the first dimension corresponding to 
landmarks, the second corresponding to X, Y, and Z, and the third corresponding to specimens.}

\item{plt}{A logical value specifying whether or not to plot logged ratios of size-scaled inter-landmark 
distances.  Defaults to \code{FALSE}.}

\item{highlight}{An integer value specifying the number of a landmark to highlight in plots. If 
provided, all logged ratios for interlandmark distances involving that landmark will be highlighted 
in the plot.  Defaults to \code{NULL}.}

\item{pt.bg}{A color value to use in plotting logged ratios.}

\item{hl.pch}{A color value to use for highlighted points when a value is supplied to \code{highlight}.}
}
\value{
A data frame providing the standard deviation of logged ratios associated with each landmark.
}
\description{
Calculates a measure of spatial variability for each landmark in relation to all other landmarks 
  for a set of objects.  Builds on the visualization method of Cole and Richtsmeier (1998), but 
  extends it for more than two specimens and adds an explicit quantification component.
  for all objects (calculated using \code{shapediff}).
}
\examples{
results <- lmkdisper(guenons$rawcoords[1:20,,guenons$species=="patas"], plt=TRUE, highlight=1)
results
}
