% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDMA_0.1_functions_analyses_2025-11-11.r
\name{meanform}
\alias{meanform}
\title{Estimate Mean Form and Landmark Covariance}
\usage{
meanform(x)
}
\arguments{
\item{x}{An array of landmarks for multiple specimens, with the first dimension corresponding to 
landmarks, the second corresponding to X, Y, and Z, and the third corresponding to specimens.}
}
\value{
A data frame providing the standard deviation of logged ratios associated with each landmark.
}
\description{
Estimates the mean form, mean landmark set, and covariance structure of lanmarks for a set of objects  
  following the algorithms provided in Lele and Richstmeier (2000).
}
\examples{
FMmean <- meanform(guenons$rawcoords[,,guenons$species=="patas"])
str(FMmean)
plot(FMmean)
}
