% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/EDMA_0.1_functions_plot_print_summary_2025-11-06.r
\name{plot.FM}
\alias{plot.FM}
\title{Plot \code{FM} Object}
\usage{
\method{plot}{FM}(
  x,
  segs = NULL,
  tris = NULL,
  pt.cols = "blue",
  pt.size = 0.5,
  pt.alpha = 0.5,
  label = FALSE,
  label.cex = 0.7
)
}
\description{
Plot form matrix.  Uses \code{cmdscale} to reconstruct landmarks for plotting.
}
\examples{
plot_lmks(guenons$rawcoords[,,guenons$species=="patas"][,,1],
          segs=guenons$wireframe_segments,
		     tris=guenons$triangles,
		     label=TRUE)
FM1 <- calcFM(guenons$rawcoords[,,guenons$species=="patas"][,,1])
str(FM1)
plot(FM1,
     segs=guenons$wireframe_segments,
		tris=guenons$triangles,
		label=TRUE)
}
