% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/EDMA_0.1_functions_plot_print_summary_2025-11-18.r
\name{plot.FMmean}
\alias{plot.FMmean}
\title{Plot \code{FMmean} Object}
\usage{
\method{plot}{FMmean}(
  x,
  segs = NULL,
  tris = NULL,
  pt.cols = "blue",
  pt.col.scale = TRUE,
  pt.size = 0.5,
  pt.size.scale = TRUE,
  pt.alpha = 0.5,
  label = FALSE,
  label.cex = 0.7,
  label.adj = 1,
  ...
)
}
\arguments{
\item{x}{A matrix of landmarks where rows correspond to landmarks and columns correspond to X, Y, and Z.}

\item{segs}{...}

\item{tris}{...}

\item{pt.cols}{...}

\item{pt.col.scale}{...}

\item{pt.size}{...}

\item{pt.size.scale}{...}

\item{pt.alpha}{...}

\item{label}{...}

\item{label.cex}{...}

\item{...}{Arguments to be passed to other functions.  Not currently used.}
}
\description{
Plot the \code{FMmean} object produced by the function \code{meanform}.
}
\examples{
patasmean<- meanform(guenons$rawcoords[,,guenons$species=="patas"])
plot(patasmean)
plot(patasmean,
     segs=guenons$wireframe_segments,
		tris=guenons$triangles,
		label=TRUE, pt.alpha=0.2, pt.size=0.3)
}
