% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/EDMA_0.1_functions_plot_print_summary_2025-11-18.r
\name{plot.SM}
\alias{plot.SM}
\title{Plot \code{SM} Object}
\usage{
\method{plot}{SM}(
  x,
  segs = NULL,
  tris = NULL,
  pt.cols = "blue",
  pt.size = 0.5,
  pt.alpha = 0.5,
  label = FALSE,
  label.cex = 0.7,
  ...
)
}
\description{
Plot shape matrix.  Uses \code{cmdscale} to reconstruct landmarks for plotting.
}
\examples{
plot_lmks(guenons$rawcoords[,,guenons$species=="patas"][,,1],
          segs=guenons$wireframe_segments,
		     tris=guenons$triangles,
		     label=TRUE)
SM1 <- calcSM(guenons$rawcoords[,,guenons$species=="patas"][,,1])
str(SM1)
plot(SM1,
     segs=guenons$wireframe_segments,
		tris=guenons$triangles,
		label=TRUE)
}
