% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/EDMA_0.1_functions_plot_print_summary_2025-11-08.r
\name{plot_lmks}
\alias{plot_lmks}
\title{Plot Object Defined by Landmarks}
\usage{
plot_lmks(
  x,
  pt.cols = "blue",
  pt.size = 0.5,
  pt.alpha = 0.5,
  label = FALSE,
  label.cex = 0.7,
  segs = NULL,
  tris = NULL,
  ...
)
}
\arguments{
\item{x}{A matrix of landmarks where rows correspond to landmarks and columns correspond to X, Y, and Z.}

\item{pt.cols}{...}

\item{pt.size}{...}

\item{pt.alpha}{...}

\item{label}{...}

\item{label.cex}{...}

\item{segs}{...}

\item{tris}{...}

\item{...}{Arguments to be passed to other functions.  Not currently used.}
}
\description{
Plot a set of landmarks.
}
\examples{
plot_lmks(guenons$rawcoords[,,guenons$species=="patas"][,,30])
plot_lmks(guenons$rawcoords[,,guenons$species=="patas"][,,30],
          segs=guenons$wireframe_segments,
		     tris=guenons$triangles,
		     label=TRUE)
}
