% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDMA_0.1_functions_general_2025-11-06.r
\name{readlmk.csv}
\alias{readlmk.csv}
\title{Read Landmark CSV file}
\usage{
readlmk.csv(file = NULL)
}
\arguments{
\item{file}{A character providing the name of the file to read in.}
}
\value{
Produces a data frame preserving only the X, Y, and Z coordinates, along with any 
   landmark names that are included as row names in the file that is read in.  The name of the
   file is retained as an attribute attached to the data frame.
}
\description{
Function for reading in a landmark CSV file of XY or XYZ coordinates for a single object.
}
