% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDMA_0.1_functions_analyses_2025-11-08.r
\name{shapediff}
\alias{shapediff}
\title{Calculate Shape Difference Between Two Objects}
\usage{
shapediff(x, y)
}
\arguments{
\item{x}{A \code{FM} or \code{SM} object.}

\item{y}{A \code{FM} or \code{SM} object.}
}
\value{
A measure of shape dissimilarity, measure as the standard deviation of the natural log 
  of ratios for all interlandmark distances, where ratios are calculated as (specimen \code{x})/(specimen \code{y}).
}
\description{
Calculate the shape dissimilarity between two objects sharing a common set of landmarks using the 
  \emph{slr} metric developed by Gordon and Wood (2013).
}
\examples{
shapediff(x=calcFM(guenons$rawcoords[,,1]), y=calcFM(guenons$rawcoords[,,2]))
}
